debugging 2

+++ c/main_project/led_control/src/screen_capture_functions.c
@@ -118,11 +118,11 @@ void yuyv_to_rgb(unsigned char *yuv_buffer,unsigned char *rgb_buffer, size_t fra
       int v  = yuv_buffer[x + 3] - 128;
 
       if(y0 != 0) {
-        int byte_index = x - rgb_buffer;
-        int pixel_index = byte_index / 3;
+        size_t byte_index = x - rgb_buffer;
+        size_t pixel_index = byte_index / 3;
 
-        int x = pixel_index % WIDTH;
-        int y = pixel_index % HEIGHT;
+        size_t x = pixel_index % WIDTH;
+        size_t y = pixel_index % HEIGHT;
         printf("y0: %d x: %d y: %d\t", y0, x, y);
       }
